% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clrcov_ilr}
\alias{clrcov_ilr}
\title{evaluate clrcov when ILR(p) ~ MVN(mu,sig)}
\usage{
clrcov_ilr(tree, mu, sig)
}
\arguments{
\item{tree}{phylogenetic tree, has to have node labels}

\item{mu}{mean(preorder)}

\item{sig}{covariance (preorder)}
}
\description{
evaluate clrcov when ILR(p) ~ MVN(mu,sig)
}
