% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clrcov_sim_log}
\alias{clrcov_sim_log}
\title{clr covariance of compositions generated from LTN(mu,sigma)}
\usage{
clrcov_sim_log(mu, sigma, tree, iter, savesamp = F, dir = NULL)
}
\description{
clr covariance of compositions generated from LTN(mu,sigma)
}
