% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count2y}
\alias{count2y}
\title{count vector to y and yl}
\usage{
count2y(cnt, tree)
}
\arguments{
\item{cnt}{a row of OTU table (preorder)}

\item{tree}{phylogenetic tree}
}
\description{
count vector to y and yl
}
