% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dm_mom}
\alias{dm_mom}
\title{MoM estimate of Dirichlet Multinomial}
\usage{
dm_mom(X)
}
\arguments{
\item{X}{n samples * K categories counts}
}
\description{
Method-of-moments estimates of the DM model

X_i|q_i ~ Multi(N_i,q_i), q_i ~ Dir(v*pi), theta=1/(1+v)
}
