% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtm_sim}
\alias{dtm_sim}
\title{generate data from DTM}
\usage{
dtm_sim(nsim, tree, theta, tau, total)
}
\arguments{
\item{theta}{DTM mean, preorder}

\item{tau}{DTM dispersion parameter, preorder}
}
\description{
generate data from DTM
}
