% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_glasso.R
\name{gibbs_glasso}
\alias{gibbs_glasso}
\title{Gibbs sampler for posterior inference of LTN with sparse Gaussian graph of nodes}
\usage{
gibbs_glasso(niter, YL, Y, r = 1, s = 0.01, SEED = 1, lambda = "hp")
}
\arguments{
\item{niter}{number of Gibbs iteration}

\item{YL}{d x N matrices of y(Al)}

\item{Y}{d x N matrices of y(A)}

\item{r}{hyperparameter for shrinkage parameter lambda. lambda ~ Ga(r,s).}

\item{s}{hyperparameter for shrinkage parameter lambda. lambda ~ Ga(r,s).}

\item{SEED}{random seed for initializing the parameters}

\item{lambda}{'hp' -- the Gamma hyperprior will be used; or a real number used as a fixed lambda, and r and s will be ignored}
}
\description{
Gibbs sampler for posterior inference of LTN with sparse Gaussian graph of nodes
}
