% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{gibbs_ltn}
\alias{gibbs_ltn}
\title{Gibbs sampler for posterior inference of LTN with sparse Gaussian graph of nodes}
\usage{
gibbs_ltn(data = NULL, Y = NULL, YL = NULL, lambda = 10, niter, SEED = 1)
}
\arguments{
\item{data}{phyloseq object containing OTU table and phylogenetic tree}

\item{Y}{(numeric) n*d matrix of y_i(A). Y and YL will be used if the phyloseq object is not provided}

\item{YL}{(numeric) n*d matrix of y_i(A_l)}

\item{lambda}{shrinkage parameter in graphical lasso prior}

\item{niter}{number of Gibbs iterations}

\item{SEED}{random seed for initializing the parameters}
}
\description{
Gibbs sampler for posterior inference of LTN with sparse Gaussian graph of nodes
}
