% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_data}
\alias{make_data}
\title{make data for two-group comparison}
\usage{
make_data(ps, test_var, test_baseline, formula_covariates, sub_var)
}
\arguments{
\item{ps}{phyloseq object}

\item{test_var}{the variable that defines the two contrasting groups}

\item{test_baseline}{the baseline group in the mixed effects model}

\item{formula_covariates}{formula of other fixed effects to be adjusted}

\item{sub_var}{random effects}
}
\description{
helper function
}
