% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeDataParser}
\alias{makeDataParser}
\title{convert phyloseq object into the format of input of the LTN-based mixed-effects model sampler}
\usage{
makeDataParser(formula, data, test_baseline)
}
\arguments{
\item{formula}{a one-sided lme4 type formula, where the first variable is the group being tested}

\item{data}{a phyloseq object with an OTU table and a tree}

\item{test_baseline}{which level of the groups to be used as baseline}
}
\description{
convert phyloseq object into the format of input of the LTN-based mixed-effects model sampler
}
