% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_pmap}
\alias{plot_pmap}
\title{plot PMAPs along the tree}
\usage{
plot_pmap(
  pmap,
  tree,
  main.text,
  alpha = NULL,
  label = NULL,
  label_nodes = NULL,
  tip_label = NULL
)
}
\arguments{
\item{pmap}{vector of PMAPs}

\item{tree}{phylogenetic tree}

\item{main.text}{title}

\item{alpha}{(optional) posterior mean of alpha}

\item{label}{labels of internal nodes}

\item{label_nodes}{a set of k nodes to be labelled by A1, ..., Ak}
}
\description{
plot PMAPs along the tree
}
