% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_timepoints}
\alias{plot_timepoints}
\title{time series plot}
\usage{
plot_timepoints(data, x, y, color, shape, facet_by)
}
\arguments{
\item{data}{data.frame}

\item{facet_by}{subject-level binary variable}
}
\description{
time series plot
}
