% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seqtab2y}
\alias{seqtab2y}
\title{transform OTU counts to node counts (without nodemat, not very effective)}
\usage{
seqtab2y(seqtab, tree, preorder = F)
}
\arguments{
\item{seqtab}{OTU table}

\item{tree}{-- phylogenetic tree}

\item{preorder}{-- whether the OTUs are ordered according to the preorder traversal of the tree}
}
\value{
Y -- {y(A): A is internal node}, nodes are ordered according to the preorder traversal of the tree

YL -- {y(A_l): A is internal node}, nodes are ordered according to the preorder traversal of the tree
}
\description{
transform OTU counts to node counts (without nodemat, not very effective)
}
