% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tax_tree}
\alias{tax_tree}
\title{build phylogenetic tree based on taxonomy table}
\usage{
tax_tree(taxtab)
}
\arguments{
\item{taxtab}{taxonomy table}
}
\value{
a rooted full binary tree (some node labels are arbitrarily assigned)
}
\description{
build phylogenetic tree based on taxonomy table
}
