% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clrcov_dtm_sim_log}
\alias{clrcov_dtm_sim_log}
\title{clr covariance of compositions generated from DTM}
\usage{
clrcov_dtm_sim_log(nsim, tree, theta, tau, SSS = 1, savesamp = F, dir = NULL)
}
\description{
clr covariance of compositions generated from DTM
}
