% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count2y}
\alias{count2y}
\title{transform count vector to y(A) and y(Al)}
\usage{
count2y(cnt, tree)
}
\arguments{
\item{cnt}{a row of OTU table (preorder)}

\item{tree}{phylogenetic tree}
}
\description{
transform count vector to y(A) and y(Al)
}
