% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dtm_sim}
\alias{dtm_sim}
\title{generate data from DTM}
\usage{
dtm_sim(nsim, tree, theta, tau, total)

dtm_sim(nsim, tree, theta, tau, total)
}
\arguments{
\item{nsim}{number of samples}

\item{tree}{DT tree}

\item{theta}{DT mean}

\item{tau}{DT dispersion parameter}

\item{total}{total counts in each sample}
}
\description{
generate data from DTM

sample from DTM
}
