% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed_effects.R
\name{gibbs_crossgroup}
\alias{gibbs_crossgroup}
\title{Gibbs sampler for cross-group comparison}
\usage{
gibbs_crossgroup(
  N,
  p,
  g = NULL,
  r,
  YL,
  Y,
  Xtest,
  Xadjust = NULL,
  grouplabel = NULL,
  c0 = 1,
  d0 = 0.001,
  c1 = 1,
  d1 = 0.001,
  nu = 3,
  niter,
  adjust = F,
  reff = F,
  reffcov = 1,
  SEED = 1,
  save_alpha_only = F,
  gprior_m = 1,
  pnull = 0.5,
  a1 = 3,
  a2 = 4,
  a1a2 = "hp",
  lambda_fixed
)
}
\arguments{
\item{Y}{-- (numeric) N*p, pre-order}
}
\description{
Gibbs sampler for cross-group comparison
}
