% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltnme.R
\name{ltnme}
\alias{ltnme}
\title{Gibbs sampler for LTN-based mixed-effects model}
\usage{
ltnme(
  formula,
  data,
  test_baseline,
  niter,
  reffcov = 2,
  SEED = 1,
  save_alpha_only = F,
  gprior_m = 100,
  pnull = 0.5,
  lambda = 10
)
}
\arguments{
\item{formula}{a lme4 type formula where the first fixed effect is the group indicator}

\item{data}{a phyloseq object}

\item{test_baseline}{the group to be used as baseline in the linear model}

\item{niter}{number of Gibbs iterations}

\item{reffcov}{specification of the precision matrix of random effects. 1: diagonal, 2: sparse with graphical lasso prior}

\item{SEED}{random seed for initializing the parameters}

\item{save_alpha_only}{whether to only return posterior samples of alpha(A)}

\item{gprior_m}{hyperparameter in prior of beta with default value 100}

\item{pnull}{prior probability of the joint null hypothesis with default value 0.5}

\item{lambda}{shrinkage parameter in the graphical lasso prior with default value 10}
}
\description{
a wrapper for \code{gibbs_crossgroup}
}
