% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltnme.R
\name{ltnme}
\alias{ltnme}
\title{LTN-based mixed-effects model}
\usage{
ltnme(
  formula,
  data,
  test_baseline,
  niter,
  reffcov = 2,
  SEED = 1,
  pi_only = F,
  gprior_m = 100,
  pnull = 0.5,
  lambda = 10
)
}
\arguments{
\item{reffcov}{1: diagonal, 2: sparse}
}
\description{
LTN-based mixed-effects model
}
